+++ counted_float/benchmarking/__init__.py
@@ -1,6 +1,7 @@
-from counted_float._core.benchmarking import FlopsBenchmarkResults, run_flops_benchmark
+from counted_float._core.benchmarking import FlopsBenchmarkResults, run_counted_float_benchmark, run_flops_benchmark
 
 __all__ = [
     "FlopsBenchmarkResults",
+    "run_counted_float_benchmark",
     "run_flops_benchmark",
 ]